moved initchain code to own file
